requirements.txt
pandas==2.1.4
pycoingecko==3.1.0
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
numpy==1.26.2
openpyxl==3.1.2
aiohttp==3.9.1
python-telegram-bot==20.7
psutil==5.9.7
